Move strip-html-tags to app/utils
