+++ app/utils/strip-html-tags.ts
